requirements.txt
django
django-rest-framework
redis
pycurl